pyproject.toml
[tool.poetry]
name = "txacct"
version = "0.1.0"
description = ""
authors = ["Rene Cunningham <rene@compounddata.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.3.2"
Flask-SQLAlchemy = "^3.0.5"
psycopg = {extras = ["c"], version = "^3.1.9"}
Flask-Migrate = "^4.0.4"
click = "^8.1.6"
requests = "^2.31.0"
nltk = "^3.8.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
openpyxl = "^3.1.2"
pytest-postgresql = "^5.0.0"
psycopg2 = "^2.9.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]
